Fix pass: the bounds test's measured line prints on newlib
%lu with a cast instead of %zu: newlib's printf on the QEMU legs does not parse %zu and shifted the two doubles that followed it, so the M55 log showed the pin under "worst" and the measurement under "pin". The values themselves were never affected (the EXPECT reads the variables); the line exists so the CMSIS and vDSP legs' numbers are in the log, so it has to read right there. M55 now prints CMSIS 1.192e-07 at N = 32 and 2.086e-07 at 4096, the named split-radix engine 5.215e-08 at 16 and 2.086e-07 at 8192, all under the 2.864e-06 pin.

+++ tests/test_fft_engine.cpp
@@ -262,8 +262,10 @@ namespace {
         for (std::size_t i = 0; i < n; ++i) {
             worst = std::fmax(worst, std::fabs(back[i] - x[i]));
         }
-        std::printf("[ measured ] ConstructsAtTheRangeBounds: %s N=%zu worst |back - x| = %.3e (pin %.3e)\n",
-                    typeid(typename Fft::engine).name(), n, static_cast<double>(worst),
+        // %lu with a cast rather than %zu: newlib's printf on the QEMU legs
+        // does not parse %zu and shifts the arguments that follow it.
+        std::printf("[ measured ] ConstructsAtTheRangeBounds: %s N=%lu worst |back - x| = %.3e (pin %.3e)\n",
+                    typeid(typename Fft::engine).name(), static_cast<unsigned long>(n), static_cast<double>(worst),
                     static_cast<double>(k_round_trip_pin));
         EXPECT_LT(worst, k_round_trip_pin) << "n=" << n;
     }
